v1

+++ cash_machine/cash_machine/models/item.py
@@ -1,6 +1,6 @@
 from sqlalchemy import CheckConstraint
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy.dialects.postgresql import TEXT, INTEGER
+from sqlalchemy.dialects.postgresql import TEXT, FLOAT, INTEGER
 from .base import Base
 
 
@@ -8,4 +8,6 @@ class Item(Base):
     __tablename__ = "items"
     id: Mapped[int] = mapped_column(INTEGER(), primary_key=True, autoincrement=True)
     title: Mapped[str] = mapped_column(TEXT())
-    price: Mapped[int] = mapped_column(INTEGER(), CheckConstraint('price > 0', name='check_price_positive'))
+    price: Mapped[int] = mapped_column(
+        FLOAT(), CheckConstraint("price > 0", name="check_price_positive")
+    )
